chore: init refactor
