+++ frontend/src/types/filter/Filter.ts
@@ -14,10 +14,6 @@ export type CourseFilter = {
     years: string[];
 } & Filter;
 
-export type PrivateCourseFilter = {
-    invitationLink: string;
-} & Filter;
-
 export type DockerImageFilter = {
     id: string;
     name: string;
@@ -64,7 +60,6 @@ export function getUserFilters(query: LocationQuery): UserFilter {
 export function getCourseFilters(query: LocationQuery): CourseFilter {
     const filters: CourseFilter = {
         search: query.search?.toString() ?? '',
-        invitationLink: 'none',
         faculties: [],
         years: [],
     };
@@ -76,20 +71,6 @@ export function getCourseFilters(query: LocationQuery): CourseFilter {
     if (query.years !== undefined) {
         filters.years = getQueryList(query.years as string | string[]);
     }
-
-    return filters;
-}
-
-/**
- * Get the private course filters from the query.
- *
- * @param query
- */
-export function getPrivateCourseFilters(query: LocationQuery): PrivateCourseFilter {
-    const filters: PrivateCourseFilter = {
-        search: query.search?.toString() ?? '',
-        invitationLink: query.invitationLink?.toString() ?? '',
-    };
 
     return filters;
 }
